penumbra(storage): implement `SubstoreStorage`
